docs/legacy_planejamento/tools/requirements-golden.txt
Pillow>=9.0.0
numpy>=1.22
scikit-image>=0.19
matplotlib>=3.5
